+++ app/domain/authentication/authn_k8s/k8s_object_lookup.rb
@@ -60,7 +60,7 @@ module Authentication
             cert_store: @cert_store,
             verify_ssl: OpenSSL::SSL::VERIFY_PEER
           },
-          http_proxy_uri: ENV['https_proxy'] || ENV['http_proxy']
+          http_proxy_uri: URI.parse(api_url).find_proxy
         }
       end